Changes of variables types

+++ src/test/java/com/leroy/magportal/ui/models/orders/ContentProductCardData.java
@@ -10,10 +10,10 @@ public class ContentProductCardData extends ToGiveAwayProductCardData {
     private String reasonForNonGiveaway = null;
 
 
-    public void increaseCreatedQuantity(int val) { this.createdQuantity += val;
+    public void increaseCreatedQuantity(Double val) { this.createdQuantity += val;
     }
 
-    public void decreaseCreatedQuantity(int val) {
+    public void decreaseCreatedQuantity(Double val) {
         this.createdQuantity -= val;
     }
